Make release.nix quickly eval to a derivation defining a nix file
This way, we can eval release.nix without too much work going on. Building the
resulting derivation will take time, but that's fine since it's at build time,
and it has nicely laid out dependencies, etc. The result is a .nix file which
defines nix-eval builds for various versions of nixpkgs and GHC. We can use this
as a build input for Hydra, which will then build all of these versions.